A first draft adding "session" support
A better name is probably anonymous sessions. The idea is you can
create an anonymous session and associate events with that
session. When you finally it is a user you recognize, you can update
the session and subsequent events.

This adds an `AddSession` function to create a new session, key
generator function, and a session event logger that works like the
normal event logger.